Move group coverage test after dependency orders changed
Move it closer to where we need to test it (config package doesn't bring
those dependencies into scope anymore).
